implementando Services

+++ REST/ProductsAPI/src/main/java/furb/web2/Dao/Item/ItemDao.java
@@ -5,8 +5,7 @@ import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 import furb.web2.Models.Item.Item;
 
-public interface ItemDao extends CrudRepository<Item, Integer> {
+public interface ItemDao extends CrudRepository<Item, Long> {
 	List<Item> findByOrderId(long orderId);
-	
 	List<Item> findByProductId(long productId);
 }
