+++ REST/ProductsAPI/src/main/java/furb/web2/Dao/Order/OrderDao.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 import furb.web2.Models.Order.Order;
 
-public interface OrderDao extends CrudRepository<Order, Integer> {
+public interface OrderDao extends CrudRepository<Order, Long> {
 	List<Order> findByUserId(long userId);
 }
