+++ REST/ProductsAPI/src/main/java/furb/web2/Dao/User/UserDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import furb.web2.Models.User.User;
 
-public interface UserDao extends CrudRepository<User, Integer> {
+public interface UserDao extends CrudRepository<User, Long> {
 	User findByUsername(String username);
 }
